Stab at memory layer for int -> long on node,rel,prop ids. Still have to fix IntArray and store layer.

+++ enterprise/kernel/src/main/java/org/neo4j/kernel/impl/core/DefaultRelationshipTypeCreator.java
@@ -94,7 +94,7 @@ public class DefaultRelationshipTypeCreator implements RelationshipTypeCreator
             try
             {
                 txManager.begin();
-                id = idGenerator.nextId( RelationshipType.class );
+                id = (int) idGenerator.nextId( RelationshipType.class );
                 persistence.createRelationshipType( id, name );
                 txManager.commit();
                 success = true;
